requirements.txt
certifi==2018.8.13
chardet==3.0.4
defusedxml==0.5.0
Django==2.1
django-recaptcha==1.4.0
idna==2.7
mysqlclient==1.3.13
oauthlib==2.1.0
pkg-resources==0.0.0
PyJWT==1.6.4
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23

